pyproject.toml
[tool.poetry]
name = "tracts"
version = "0.1.0"
description = ""
authors = ["Andreas Atle <atle.andreas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytesseract = "^0.3.13"
pillow = "^10.4.0"
argparse = "^1.4.0"
pdf2image = "^1.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
